Add first implementation of asynchronous inference
The used approach relies on a chain of Promises to synchronize and make
the inference asynchronous, but still sequential (so notes are still in
order and internal state is managed correctly).

Synchronous and asynchronous execution must not be mixed!
All asynchronous calls must have settled before the next synchronous
inference is started!

Note that this implementation is just for testing if asynchronous
inference reduces audible pops in audio playback (which it does not).
It is mostly a hack on top of the synchrous PianoGenie code.

Do not use this in production!